+++ atom/model_ops/dcp_ops.py
@@ -8,6 +8,8 @@ across DCP ranks using LSE (Log-Sum-Exp) correction.
 Uses vllm-style algorithm: AllGather LSE -> correct local output -> ReduceScatter.
 """
 
+from typing import Optional
+
 import numpy as np
 import torch
 import triton
@@ -488,3 +490,427 @@ def dcp_pack_topk_candidates(
         torch.full_like(local_idx, -1),
     )
     out_pair.view(torch.int32)[1].copy_(gid)
+
+
+# ---------------------------------------------------------------------------
+# DCP sparse index filter + round-robin localize (decode & prefill).
+# Two-pass compacting filter: count this rank's owned top-k, then pack the
+# owned slots to the front of each region (no -1 holes -- see cp_lse_ag_out_rs).
+# ---------------------------------------------------------------------------
+
+
+@triton.jit
+def _count_owned_dcp_kernel(
+    qo_indptr,  # int32 [num_requests + 1]
+    global_kv_indptr,  # int32 [num_requests + 1] -- GLOBAL context (column range)
+    token_indices_ptr,  # int32 [num_tokens, NUM_TOPK_TOKENS] -- GLOBAL top-k positions
+    out_counts,  # int32 [num_requests] -- owned top-k count per request
+    DCP_RANK: tl.constexpr,
+    DCP_WORLD: tl.constexpr,
+    NUM_TOPK_TOKENS: tl.constexpr,
+    BLOCK_N: tl.constexpr,
+    ti_stride0,
+    ti_stride1,
+):
+    """Pass 1 of the compacting DCP filter: how many of the global top-k
+    positions does this rank own, per request? Its exclusive cumsum gives the
+    compacted output offsets used by ``_compact_filter_dcp_kernel``.
+
+    qlen==1 only (DCP + sparse + MTP is rejected upstream), so each request has
+    exactly one query token.
+    """
+    batch_id = tl.program_id(0)
+    g_kv_start = tl.load(global_kv_indptr + batch_id)
+    g_kv_len = tl.load(global_kv_indptr + batch_id + 1) - g_kv_start
+    token_id = tl.load(qo_indptr + batch_id)
+
+    count = 0
+    for tile_start in range(0, NUM_TOPK_TOKENS, BLOCK_N):
+        indice_id = tile_start + tl.arange(0, BLOCK_N)
+        col_valid = (indice_id < g_kv_len) & (indice_id < NUM_TOPK_TOKENS)
+        ti_ptr = token_indices_ptr + token_id * ti_stride0 + indice_id * ti_stride1
+        tok = tl.load(ti_ptr, mask=col_valid, other=-1)
+        owned = col_valid & (tok >= 0) & ((tok % DCP_WORLD) == DCP_RANK)
+        count += tl.sum(owned.to(tl.int32))
+
+    tl.store(out_counts + batch_id, count)
+
+
+@triton.jit
+def _compact_filter_dcp_kernel(
+    qo_indptr,  # int32 [num_requests + 1]
+    global_kv_indptr,  # int32 [num_requests + 1] -- GLOBAL context (column range)
+    out_kv_indptr,  # int32 [num_requests + 1] -- COMPACTED offsets (cumsum of pass 1)
+    block_table,  # int32 [num_req, max_num_blocks_per_req] -- logical(global) blocks
+    token_indices_ptr,  # int32 [num_tokens, NUM_TOPK_TOKENS] -- GLOBAL top-k positions
+    out_kv_indices,  # int32 [>= out_kv_indptr[-1]]
+    DCP_RANK: tl.constexpr,
+    DCP_WORLD: tl.constexpr,
+    PAGE_SIZE: tl.constexpr,  # runner (physical) block size
+    NUM_TOPK_TOKENS: tl.constexpr,
+    BLOCK_N: tl.constexpr,
+    ti_stride0,
+    ti_stride1,
+    bt_stride0: tl.int64,
+    bt_stride1: tl.constexpr,
+):
+    # DCP round-robin (interleave=1): a GLOBAL position g is owned by rank
+    # ``g % W``; on the owner rank its physical slot follows the round-robin
+    # (virtual-block) layout used by _dcp_round_robin_slot / ATOM PR #847:
+    #     vbs  = PAGE_SIZE * W
+    #     slot = block_table[req, g // vbs] * PAGE_SIZE + (g % vbs) // W
+    # token_indices holds GLOBAL positions (the indexer scored the full sequence
+    # via all-gathered logits). This rank keeps ONLY the positions it owns and
+    # writes them COMPACTED to the front of its region -- no -1 holes. Holes are
+    # exactly what breaks aiter's lse path (immediate fault on the persistent
+    # kernel, silently unwritten lse on the split-KV one).
+    #
+    # Compaction is order-preserving (tl.cumsum within a tile plus a running
+    # offset across tiles) rather than atomic-allocated like vLLM, so the KV
+    # order -- and hence the floating-point accumulation order -- is
+    # deterministic run to run, which the dcp=1 vs dcp=N comparison relies on.
+    #
+    # NOTE: the slot is computed from block_table directly (like vLLM) rather
+    # than gathered from a precomputed kv_indices -- the DCP round-robin
+    # per-token slot array does not exist on the sparse path (dense reads go
+    # through block_tables in-kernel).
+    batch_id = tl.program_id(0)
+
+    g_kv_start = tl.load(global_kv_indptr + batch_id)
+    g_kv_len = tl.load(global_kv_indptr + batch_id + 1) - g_kv_start
+    out_kv_start = tl.load(out_kv_indptr + batch_id)
+    token_id = tl.load(qo_indptr + batch_id)
+
+    vbs = PAGE_SIZE * DCP_WORLD
+    written = 0
+    for tile_start in range(0, NUM_TOPK_TOKENS, BLOCK_N):
+        indice_id = tile_start + tl.arange(0, BLOCK_N)
+        # Columns within [0, min(global_ctx, topk)) carry a real top-k position.
+        col_valid = (indice_id < g_kv_len) & (indice_id < NUM_TOPK_TOKENS)
+
+        ti_ptr = token_indices_ptr + token_id * ti_stride0 + indice_id * ti_stride1
+        tok = tl.load(ti_ptr, mask=col_valid, other=-1)  # GLOBAL position
+
+        idx_valid = col_valid & (tok >= 0) & ((tok % DCP_WORLD) == DCP_RANK)
+
+        block_id = tok // vbs
+        inblock_offset = (tok % vbs) // DCP_WORLD
+        physical_block = tl.load(
+            block_table + batch_id * bt_stride0 + block_id * bt_stride1,
+            mask=idx_valid,
+            other=0,
+        )
+        slot = physical_block * PAGE_SIZE + inblock_offset
+
+        # Exclusive prefix sum of the owned mask -> destination inside this tile.
+        owned_i32 = idx_valid.to(tl.int32)
+        dst = written + tl.cumsum(owned_i32, axis=0) - owned_i32
+        tl.store(out_kv_indices + out_kv_start + dst, slot, mask=idx_valid)
+        written += tl.sum(owned_i32)
+
+
+def triton_filter_and_convert_dcp_index(
+    qo_indptr: torch.Tensor,  # int32 [num_requests + 1]
+    global_kv_indptr: torch.Tensor,  # int32 [num_requests + 1]
+    block_table: torch.Tensor,  # int32 [num_req, max_num_blocks_per_req] logical
+    token_indices: torch.Tensor,  # int32 [num_tokens, NUM_TOPK_TOKENS] GLOBAL pos
+    dcp_rank: int,
+    dcp_world_size: int,
+    block_size: int,  # runner (physical) block size == PAGE_SIZE
+    out_kv_indptr: torch.Tensor,  # int32 [num_requests + 1] COMPACTED, written here
+    owned_counts: torch.Tensor,  # int32 [>= num_requests] scratch for pass 1
+    NUM_TOPK_TOKENS: int = 2048,
+    BLOCK_N: int = 128,
+    out: Optional[torch.Tensor] = None,
+):
+    """DCP (interleave=1) filter + round-robin localize of global top-k positions,
+    **compacting** each rank's owned slots to the front of its region.
+
+    ``token_indices[token_id, indice_id]`` is a GLOBAL token position selected by
+    the indexer (scored over the full sequence via all-gathered logits). This
+    rank keeps a position ``g`` only if ``g % W == dcp_rank`` and maps it to its
+    physical slot via the round-robin (virtual-block) layout, computed directly
+    from ``block_table`` (like vLLM):
+        vbs  = block_size * W
+        slot = block_table[req, g // vbs] * block_size + (g % vbs) // W
+
+    Non-owned positions are **dropped**, not marked: the kept slots are packed
+    contiguously (original top-k order preserved) and ``out_kv_indptr`` is
+    rewritten to the resulting per-request lengths. This replaces the earlier
+    "fixed length + -1 sentinel" layout, whose holes broke aiter's lse output.
+    Because the kept count depends on the per-layer top-k selection,
+    ``out_kv_indptr`` is layer-dependent and must be recomputed on every call --
+    it cannot feed the once-per-step persistent metadata, which is why sparse+DCP
+    runs non-persistent for now.
+
+    The 8 ranks' kept sets are disjoint and their union is exactly the global
+    top-k, which is what makes the downstream ``cp_lse_ag_out_rs`` merge valid.
+    """
+    assert token_indices.dtype == torch.int32
+    assert token_indices.shape[1] == NUM_TOPK_TOKENS
+    assert NUM_TOPK_TOKENS % BLOCK_N == 0, (
+        f"NUM_TOPK_TOKENS ({NUM_TOPK_TOKENS}) must be divisible by"
+        f"BLOCK_N ({BLOCK_N})"
+    )
+    assert 0 <= dcp_rank < dcp_world_size
+    assert out is not None, "sparse_kv_indices_buffer (out) is required"
+
+    num_batch = global_kv_indptr.shape[0] - 1
+
+    qo_indptr_c = qo_indptr.contiguous()
+    global_kv_indptr_c = global_kv_indptr.contiguous()
+    block_table_c = block_table.contiguous()
+    token_indices_c = token_indices.contiguous()
+
+    ti_stride0, ti_stride1 = token_indices_c.stride()
+    bt_stride0, bt_stride1 = block_table_c.stride()
+    grid = (num_batch,)
+
+    # Pass 1: per-request count of owned top-k positions.
+    counts = owned_counts[:num_batch]
+    _count_owned_dcp_kernel[grid](
+        qo_indptr_c,
+        global_kv_indptr_c,
+        token_indices_c,
+        counts,
+        dcp_rank,
+        dcp_world_size,
+        NUM_TOPK_TOKENS,
+        BLOCK_N,
+        ti_stride0,
+        ti_stride1,
+    )
+
+    # Exclusive cumsum -> compacted offsets. Written in place so the caller's
+    # tensor (and anything already holding a view of it) sees the update.
+    # dtype=int32 keeps the accumulation in int32 (torch would promote integral
+    # cumsum to int64 by default, which the kernels' int32 pointers reject).
+    # zero_() rather than `out_kv_indptr[0] = 0`: assigning a Python scalar goes
+    # through a host->device copy, which HIP rejects while a graph is capturing
+    # (hipErrorStreamCaptureUnsupported). Everything here must stay device-side.
+    out_kv_indptr[:1].zero_()
+    torch.cumsum(counts, dim=0, dtype=torch.int32, out=out_kv_indptr[1 : num_batch + 1])
+
+    # Pass 2: write the owned slots packed to the front of each region.
+    _compact_filter_dcp_kernel[grid](
+        qo_indptr_c,
+        global_kv_indptr_c,
+        out_kv_indptr,
+        block_table_c,
+        token_indices_c,
+        out,
+        dcp_rank,
+        dcp_world_size,
+        block_size,
+        NUM_TOPK_TOKENS,
+        BLOCK_N,
+        ti_stride0,
+        ti_stride1,
+        bt_stride0,
+        bt_stride1,
+    )
+    return out
+
+
+@triton.jit
+def _count_owned_dcp_prefill_kernel(
+    dsa_kv_indptr,  # int32 [num_tokens + 1] -- GLOBAL per-token candidate counts
+    token_to_seq_idxs,  # int32 [num_tokens]
+    topk_indices,  # int32 [num_tokens, NUM_TOPK_TOKENS] -- FLAT KV indices
+    cu_seqlens_k,  # int32 [num_req + 1] -- per-seq base of the flat KV axis
+    out_counts,  # int32 [num_tokens]
+    DCP_RANK: tl.constexpr,
+    DCP_WORLD: tl.constexpr,
+    NUM_TOPK_TOKENS: tl.constexpr,
+    BLOCK_N: tl.constexpr,
+    ti_stride0: tl.int64,
+    ti_stride1: tl.constexpr,
+):
+    """Pass 1 of the compacting DCP filter for sparse PREFILL: how many of the
+    global top-k candidates does this rank own, per QUERY TOKEN?
+
+    Differs from the decode twin (`_count_owned_dcp_kernel`) in two ways:
+    the row unit is a query token rather than a request (prefill has many query
+    tokens per request), and `topk_indices` holds a FLAT index into the
+    concatenated KV of all sequences, so the within-sequence position -- which
+    is what the round-robin owner is derived from -- is `indice - cu_seqlens_k[req]`.
+    """
+    token_id = tl.program_id(0)
+    req_id = tl.load(token_to_seq_idxs + token_id)
+    kv_start = tl.load(dsa_kv_indptr + token_id)
+    kv_len = tl.load(dsa_kv_indptr + token_id + 1) - kv_start
+    base = tl.load(cu_seqlens_k + req_id)
+
+    count = 0
+    for tile_start in range(0, NUM_TOPK_TOKENS, BLOCK_N):
+        col_id = tile_start + tl.arange(0, BLOCK_N)
+        col_valid = (col_id < kv_len) & (col_id < NUM_TOPK_TOKENS)
+        indice = tl.load(
+            topk_indices + token_id * ti_stride0 + col_id * ti_stride1,
+            mask=col_valid,
+            other=-1,
+        )
+        pos = indice - base  # position within the sequence
+        owned = col_valid & (indice >= 0) & ((pos % DCP_WORLD) == DCP_RANK)
+        count += tl.sum(owned.to(tl.int32))
+
+    tl.store(out_counts + token_id, count)
+
+
+@triton.jit
+def _compact_filter_dcp_prefill_kernel(
+    dsa_kv_indptr,  # int32 [num_tokens + 1] -- GLOBAL per-token candidate counts
+    out_kv_indptr,  # int32 [num_tokens + 1] -- COMPACTED offsets (cumsum of pass 1)
+    token_to_seq_idxs,  # int32 [num_tokens]
+    topk_indices,  # int32 [num_tokens, NUM_TOPK_TOKENS] -- FLAT KV indices
+    cu_seqlens_k,  # int32 [num_req + 1]
+    block_table,  # int32 [num_req, max_num_blocks_per_req] -- logical(global) blocks
+    out_kv_indices,  # int32 [>= out_kv_indptr[-1]]
+    DCP_RANK: tl.constexpr,
+    DCP_WORLD: tl.constexpr,
+    PAGE_SIZE: tl.constexpr,  # runner (physical) block size
+    NUM_TOPK_TOKENS: tl.constexpr,
+    BLOCK_N: tl.constexpr,
+    ti_stride0: tl.int64,
+    ti_stride1: tl.constexpr,
+    bt_stride0: tl.int64,
+    bt_stride1: tl.constexpr,
+):
+    """Pass 2 for sparse PREFILL: keep only this rank's owned candidates, map
+    them through the round-robin (virtual-block) layout, and pack them to the
+    front of the token's region.
+
+        vbs  = PAGE_SIZE * W
+        slot = block_table[req, pos // vbs] * PAGE_SIZE + (pos % vbs) // W
+
+    Same rationale as the decode twin: no `-1` holes (they break aiter's lse
+    path) and compaction is order-preserving so the fp accumulation order is
+    deterministic.
+    """
+    token_id = tl.program_id(0)
+    req_id = tl.load(token_to_seq_idxs + token_id)
+    kv_start = tl.load(dsa_kv_indptr + token_id)
+    kv_len = tl.load(dsa_kv_indptr + token_id + 1) - kv_start
+    base = tl.load(cu_seqlens_k + req_id)
+    out_kv_start = tl.load(out_kv_indptr + token_id)
+
+    vbs = PAGE_SIZE * DCP_WORLD
+    written = 0
+    for tile_start in range(0, NUM_TOPK_TOKENS, BLOCK_N):
+        col_id = tile_start + tl.arange(0, BLOCK_N)
+        col_valid = (col_id < kv_len) & (col_id < NUM_TOPK_TOKENS)
+        indice = tl.load(
+            topk_indices + token_id * ti_stride0 + col_id * ti_stride1,
+            mask=col_valid,
+            other=-1,
+        )
+        pos = indice - base
+        idx_valid = col_valid & (indice >= 0) & ((pos % DCP_WORLD) == DCP_RANK)
+
+        block_id = pos // vbs
+        inblock_offset = (pos % vbs) // DCP_WORLD
+        physical_block = tl.load(
+            block_table + req_id * bt_stride0 + block_id * bt_stride1,
+            mask=idx_valid,
+            other=0,
+        )
+        slot = physical_block * PAGE_SIZE + inblock_offset
+
+        owned_i32 = idx_valid.to(tl.int32)
+        dst = written + tl.cumsum(owned_i32, axis=0) - owned_i32
+        tl.store(out_kv_indices + out_kv_start + dst, slot, mask=idx_valid)
+        written += tl.sum(owned_i32)
+
+    # A row this rank owns nothing of is left EMPTY (zero-length region). That is
+    # both legal and correct for mla_decode_fwd: it writes lse = -inf, which
+    # cp_lse_ag_out_rs turns into a zero weight. Its `o` comes out NaN (0/0), so
+    # the caller zeroes those rows -- see _forward_prefill_mla. No dummy candidate
+    # is injected; the attention never sees fabricated KV.
+
+
+def triton_filter_and_convert_dcp_index_prefill(
+    dsa_kv_indptr: torch.Tensor,  # int32 [num_tokens + 1] GLOBAL counts
+    token_to_seq_idxs: torch.Tensor,  # int32 [num_tokens]
+    topk_indices: torch.Tensor,  # int32 [num_tokens, NUM_TOPK_TOKENS] FLAT indices
+    cu_seqlens_k: torch.Tensor,  # int32 [num_req + 1]
+    block_table: torch.Tensor,  # int32 [num_req, max_num_blocks_per_req] logical
+    dcp_rank: int,
+    dcp_world_size: int,
+    block_size: int,  # runner (physical) block size == PAGE_SIZE
+    out_kv_indptr: torch.Tensor,  # int32 [num_tokens + 1] COMPACTED, written here
+    owned_counts: torch.Tensor,  # int32 [>= num_tokens] scratch for pass 1
+    NUM_TOPK_TOKENS: int = 2048,
+    BLOCK_N: int = 128,
+    out: Optional[torch.Tensor] = None,
+):
+    """Sparse-PREFILL twin of ``triton_filter_and_convert_dcp_index``.
+
+    The decode version keys on requests (qlen==1); prefill has one row per query
+    token and its ``topk_indices`` are flat KV indices rather than
+    within-sequence positions. Everything else -- two passes, in-place int32
+    cumsum, order-preserving compaction -- is identical, and the same
+    layer-scoped buffers are reused (they are sized ``max_num_batched_tokens``,
+    which bounds the prefill token count too).
+    """
+    assert topk_indices.dtype == torch.int32
+    assert topk_indices.shape[1] == NUM_TOPK_TOKENS
+    assert NUM_TOPK_TOKENS % BLOCK_N == 0, (
+        f"NUM_TOPK_TOKENS ({NUM_TOPK_TOKENS}) must be divisible by"
+        f"BLOCK_N ({BLOCK_N})"
+    )
+    assert 0 <= dcp_rank < dcp_world_size
+    assert out is not None, "sparse_kv_indices_buffer (out) is required"
+
+    num_tokens = dsa_kv_indptr.shape[0] - 1
+
+    dsa_kv_indptr_c = dsa_kv_indptr.contiguous()
+    token_to_seq_idxs_c = token_to_seq_idxs.contiguous()
+    topk_indices_c = topk_indices.contiguous()
+    cu_seqlens_k_c = cu_seqlens_k.contiguous()
+    block_table_c = block_table.contiguous()
+
+    ti_stride0, ti_stride1 = topk_indices_c.stride()
+    bt_stride0, bt_stride1 = block_table_c.stride()
+    grid = (num_tokens,)
+
+    counts = owned_counts[:num_tokens]
+    _count_owned_dcp_prefill_kernel[grid](
+        dsa_kv_indptr_c,
+        token_to_seq_idxs_c,
+        topk_indices_c,
+        cu_seqlens_k_c,
+        counts,
+        dcp_rank,
+        dcp_world_size,
+        NUM_TOPK_TOKENS,
+        BLOCK_N,
+        ti_stride0,
+        ti_stride1,
+    )
+
+    # Same device-side-only cumsum as the decode path (see its comment for why
+    # zero_() and dtype=torch.int32 are required).
+    out_kv_indptr[:1].zero_()
+    torch.cumsum(
+        counts, dim=0, dtype=torch.int32, out=out_kv_indptr[1 : num_tokens + 1]
+    )
+
+    _compact_filter_dcp_prefill_kernel[grid](
+        dsa_kv_indptr_c,
+        out_kv_indptr,
+        token_to_seq_idxs_c,
+        topk_indices_c,
+        cu_seqlens_k_c,
+        block_table_c,
+        out,
+        dcp_rank,
+        dcp_world_size,
+        block_size,
+        NUM_TOPK_TOKENS,
+        BLOCK_N,
+        ti_stride0,
+        ti_stride1,
+        bt_stride0,
+        bt_stride1,
+    )
+    return out
